pyproject.toml
[project]
name = "aura-frame-manager"
version = "0.1.0"
description = "simple scripts to control aura frames"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]
